pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "heater_amd_controller"
version = "1.0"
requires-python = ">=3.12"

dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "psutil>=7.1.1",
    "pyserial>=3.5",
    "pyvisa>=1.15.0",
    "pyvisa-py>=0.8.1",
    "zeroconf>=0.148.0",
]

[project.optional-dependencies]
dev = ["ruff", "ty"]

[tool.hatch.build.targets.wheel]
packages = ["src/heater_amd_controller"]


[tool.ruff]
target-version = "py312"

line-length = 100

# チェックするエラー
select = ["ALL"]
# 除外するエラー
ignore = [
    "ANN002", # missing-type-args
    "ANN003", # missing-type-kwargs
    "COM812", # 末尾のカンマを必須としない
    "INP001", # __init__.pyを必須としない
    "T20",    # flake8-print
    "D1",     # docstring除外する
    "EXE",    # flake8-executable
    "D400",
    "D415",   # docstringの末尾ピリオドの強制を無効化
    "FBT001",
    "FBT002", # 関数の引数にbooleanを許可
    "ERA001",
]
# 警告はするが、自動修正はしないエラー
unfixable = [
    "F401",   # unused-import : 使用していないモジュールの自動削除停止
    "F841",   # unused-variable : 使用していない変数の自動削除停止
    "ERA001", # commented-out-code : コメントアウトしたpythonコードのエラー
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-third-party = ["fastapi", "pydantic", "starlette"]

[tool.ruff.format]
quote-style = "double"


[tool.ruff.pylint]
max-args = 10
